Move SpanFactory to correct package.
